"add new faculty" updated for faculty settings

+++ fyp/fulltime/alloc/submit_savewl.php
@@ -46,17 +46,17 @@
 			
 			$exemption = GetSQLValueString(trim($_REQUEST['exemption_'.$staffid]),"int");
 			if($exemption === "NULL") $exemption = 0;
-			$name = GetSQLValueString(trim($_REQUEST['name_' .$staffid]),"text");
-			$name2 = GetSQLValueString(trim($_REQUEST['name2_' .$staffid]),"text");
-			$email = str_replace(' ' , '',$_REQUEST['email_' .$staffid]);
+			$name = GetSQLValueString(strtolower(trim($_REQUEST['name_' .$staffid])),"text");
+			$name2 = GetSQLValueString(strtolower(trim($_REQUEST['name2_' .$staffid])),"text");
+			$email = strtolower(trim($_REQUEST['email_' .$staffid]));
 			$canExamine = isset($_REQUEST['examine_'.$staffid]);
 			$id = explode("@",$email)[0];
 
 			// update staff id when email address is edited.
-			if ($email != str_replace(" ","",$curStaff['email'])) {
+			if ($email != strtolower(trim($curStaff['email']))) {
 
-				$Stmt = sprintf("UPDATE %s SET id = '%s', email='%s' where id ='%s'", $TABLES["staff"], $id,$email,$curStaff['id']);
-				$DBOBJ_Result = $conn_db_ntu->prepare($Stmt);
+				$query_Update = sprintf("UPDATE %s SET id = '%s', email='%s' where id ='%s'", $TABLES["staff"], $id,$email,$curStaff['id']);
+				$DBOBJ_Result = $conn_db_ntu->prepare($query_Update);
 				$DBOBJ_Result->execute();
 			}
 
@@ -67,9 +67,9 @@
 			$values[] = sprintf("(%s, %s, %s, %s, %d, %d)",
 								$id,$email, $name, $name2, $exemption, $canExamine);
 
-			if (trim($_REQUEST['name_' .$staffid]) != $curStaff['name'] || trim($_REQUEST['name2_' .$staffid]) != $curStaff['name2'] || $canExamine != $curStaff['examine'] || $exemption != $curStaff['exemption']) {
-				$Stmt = sprintf("UPDATE %s SET name=%s, name2=%s, examine=%d, exemption=%d where id =%s", $TABLES["staff"],$name,$name2,$canExamine,$exemption,$id);
-				$DBOBJ_Result = $conn_db_ntu->prepare($Stmt);
+			if (strtolower(trim($_REQUEST['name_' .$staffid])) != $curStaff['name'] || strtolower(trim($_REQUEST['name2_' .$staffid])) != $curStaff['name2'] || $canExamine != $curStaff['examine'] || $exemption != $curStaff['exemption']) {
+				$query_Update = sprintf("UPDATE %s SET name=%s, name2=%s, examine=%d, exemption=%d where id =%s", $TABLES["staff"],$name,$name2,$canExamine,$exemption,$id);
+				$DBOBJ_Result = $conn_db_ntu->prepare($query_Update);
 				$DBOBJ_Result->execute();
 			}
 
@@ -81,52 +81,39 @@
 		}
 	}
 
-	if (!empty($delete)) {
+	if (!empty($delete) && isset($delete)) {
 		$query_Delete = sprintf("DELETE FROM %s WHERE %s", $TABLES["staff"], implode(" OR ", $delete));
-		$conn_db_ntu->exec($query_Delete);
+		$DBOBJ_Result = $conn_db_ntu->prepare($query_Delete);
+		$DBOBJ_Result->execute();
 
 	}
 
 
+	$c = 0;
 	if (isset($_REQUEST['newEmail'])) {
-		echo "ccccc";
+		foreach ($_REQUEST['newEmail'] as $newRecord) {
+			$newID = GetSQLValueString(explode("@",strtolower(trim($_REQUEST['newEmail'][$c])))[0],"text");
+			$newEmail = GetSQLValueString(strtolower($_REQUEST['newEmail'][$c]),"text");
+			$newName = GetSQLValueString(strtolower(trim($_REQUEST['newName'][$c])),"text");
+			$newName2 = GetSQLValueString(strtolower(trim($_REQUEST['newName2'][$c])),"text");
+			$newExemption = GetSQLValueString(trim($_REQUEST['newExemption'][$c]),"int");
+			$newExamine = isset($_REQUEST['newExamine'][$c]);
+
+
+			$query_Insert = sprintf("INSERT INTO %s (id, email, name, name2, exemption, examine) VALUES (%s, %s, %s, %s, %d, %d) 
+		    ON DUPLICATE KEY UPDATE name=%s, name2=%s, exemption=%d, examine=%d ", $TABLES["staff"],$newID,$newEmail,$newName,$newName2,$newExemption,$newExamine,$newName,$newName2,$newExemption,$newExamine);
+			$DBOBJ_Result = $conn_db_ntu->prepare($query_Insert);
+			$DBOBJ_Result->execute();
+
+			$c++;
+		}
 	}
 
-	//$query_UpdateExemption = sprintf("UPDATE %s SET email=%s, name=%s, name2=%s, exemption=%d, examine=%d WHERE id=%s", $TABLES["staff"], )
-
-	//adding new staff.
-	/*$c = 1;
-	while (isset($_REQUEST['newStaffID_'. strval($c)])) {
-		$name = GetSQLValueString(trim($_REQUEST['newName_' . strval($c)]),"text");
-		$name2 = GetSQLValueString(trim($_REQUEST['newName2_' . strval($c)]),"text");
-		$email = str_replace(' ' , '',$_REQUEST['newEmail_' . strval($c)]);
-		$canExamine = isset($_REQUEST['newExamine_'. strval($c)]);
-		$id = explode("@",$email)[0];
-		//$id = GetSQLValueString(trim($_REQUEST['newStaffID_' . strval($c)]),"text");
-		$exemption = GetSQLValueString(trim($_REQUEST['newExemption_'. strval($c)]),"int");
-		if($exemption === "NULL") $exemption = 0;
-
-		$email = GetSQLValueString($email,"text");
-		$id = GetSQLValueString($id,"text");
-
-
-		$values[] = sprintf("(%s, %s, %s, %s, %d, %d)",
-			$id,$email, $name, $name2, $exemption, $canExamine);
-
-		$c++;
-	}
-	*/
-
-	
-	/*$query_UpdateExemption  = sprintf("INSERT INTO %s (`id`,`email`,`name`,`name2`,`exemption`,`examine`) VALUES %s
-ON DUPLICATE KEY UPDATE `email`=VALUES(`email`),`name`=VALUES(`name`),`name2`=VALUES(`name2`),`exemption`=VALUES(`exemption`), `examine`=VALUES(`examine`)",
-							$TABLES['staff'],
-							implode(",", $values));
-
-	$conn_db_ntu->exec($query_UpdateExemption);*/
 	$conn_db_ntu = null;
 	unset($rsStaff);
 	unset($values);
+	unset($delete);
+	unset($c);
 
 	if(isset ($_REQUEST['validate'])) {
 		   header("location:examiner_setting.php?validate=1");
